moved join_estimates_to_ascwds function to utils
