+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/utils/utils.py
@@ -1,5 +1,5 @@
 import polars as pl
-
+from typing import Final
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 from utils.column_values.categorical_column_values import (
     EstimateFilledPostsSource,
@@ -8,6 +8,57 @@ from utils.column_values.categorical_column_values import (
 from utils.value_labels.ascwds_worker.ascwds_worker_jobgroup_dictionary import (
     AscwdsWorkerValueLabelsJobGroup,
 )
+
+ROW_ID: Final[str] = "id"
+
+JobRoleEnumType = pl.Enum(AscwdsWorkerValueLabelsJobGroup.all_roles())
+
+
+def join_estimates_to_ascwds(
+    estimates_lf: pl.LazyFrame,
+    ascwds_lf: pl.LazyFrame,
+) -> pl.LazyFrame:
+    """Join job role estimates to ASCWDS counts ensuring a row for every job role.
+
+    Performs a cross join on the estimates join keys first to ensure there is a row for
+    every job role across all time periods for each location. This is then joined with
+    the ASCWDS data.
+
+    Args:
+        estimates_lf (pl.LazyFrame): Input LazyFrame with Estimates data.
+        ascwds_lf (pl.LazyFrame): Input LazyFrame with ASC-WDS job role counts.
+
+    Returns:
+        pl.LazyFrame: Joined LazyFrame with a row for every job role.
+
+    """
+    join_keys = [
+        IndCQC.ascwds_workplace_import_date,
+        IndCQC.establishment_id,
+    ]
+    job_role_labels = IndCQC.main_job_role_clean_labelled
+
+    narrow_keys_lf = estimates_lf.select([ROW_ID] + join_keys)
+
+    roles_lf = pl.LazyFrame(
+        data=[(role,) for role in AscwdsWorkerValueLabelsJobGroup.all_roles()],
+        schema={job_role_labels: JobRoleEnumType},
+        orient="row",
+    )
+
+    expanded_keys_lf = narrow_keys_lf.join(roles_lf, how="cross")
+
+    expanded_counts_lf = expanded_keys_lf.join(
+        other=ascwds_lf,
+        on=join_keys + [job_role_labels],
+        how="left",
+    )
+
+    return estimates_lf.join(
+        expanded_counts_lf.drop(join_keys),
+        on=ROW_ID,
+        how="right",
+    ).drop(join_keys)
 
 
 def nullify_job_role_count_when_source_not_ascwds(lf: pl.LazyFrame) -> pl.LazyFrame:
